api/requirements.txt
flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
flask-cors==4.0.0
pyjwt==2.8.0
bcrypt==4.1.1
werkzeug==3.0.1
google-generativeai==0.3.1

